+++ tests/Jobs/BaseJobTest.php
@@ -201,7 +201,18 @@ class BaseJobTest extends QlessTestCase
         $this->assertTrue($job->ttl() > $before);
     }
 
-    // Times out the job now rather than when its lock is normally set to expire
+    /** @test */
+    public function shouldStartTrackingJob()
+    {
+        $this->client->queues['foo']->put('Foo', [], 'jid');
+        $this->assertFalse($this->client->jobs['jid']->tracked);
+
+        $this->client->jobs['jid']->track();
+        $this->assertTrue($this->client->jobs['jid']->tracked);
+
+        $this->client->jobs['jid']->untrack();
+        $this->assertFalse($this->client->jobs['jid']->tracked);
+    }
 
     /**
      * @test
@@ -423,6 +434,7 @@ class BaseJobTest extends QlessTestCase
             ['expires', 'double'], // I ❤︎ PHP
             ['remaining', 'integer'],
             ['retries', 'integer'],
+            ['tracked', 'boolean'],
             ['description', 'string'],
         ];
     }