+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added basic event system
 - Added initial `qlessd` daemon
 - Added ability to adjust a job's priority while it's still waiting in a queue
+- Added `Qless\Jobs\Collection::tagged` to fetches a list of tagged JIDs associated with provided tag
+- Added `Qless\Jobs\BaseJob::track` and `Qless\Jobs\BaseJob::untrack` to (un)flagging a job as important
 
 ### Changed
 - PHP 5.x no longer supported. Minimal required version is 7.1